[move/#5] Route 코드를 Screen 파일로 이동
